Changed location of isValid
