+++ src/validation/index.js
@@ -1,9 +1,28 @@
 import 'source-map-support/register';
 
 import chalk from 'chalk';
-import { isPlainObject } from 'lodash';
+import { isPlainObject, isFunction, isRegExp } from 'lodash';
 
-import { isValid } from './helpers';
+/**
+ * Helper to use a validator.
+ *
+ * @param {object} value - Something to validate
+ * @param {function|RegExp} validator - A validator
+ * @return {boolean} Returns if valid or not
+ */
+export function isValid(value, validator) {
+    if (isFunction(validator)) {
+        return validator(value);
+    } else if (isRegExp(validator)) {
+        if (!validator.test(value.toString())) {
+            return 'Did not match the regexp: ' + validator;
+        }
+
+        return true;
+    }
+
+    throw new Error('Structure of configuration does not align with validation.');
+}
 
 /**
  * Validates the provided configuration object
